+++ .eslintrc
@@ -43,6 +43,7 @@
         "allowExpressionValues": true
       }
     ],
-    "react/no-unknown-property": ["error", { "ignore": ["intensity", "position", "rotation", "dispose", "rotation", "geometry", "material"] }]
+    "react/no-unknown-property": ["error", { "ignore": ["intensity", "position", "rotation", "dispose", "rotation", "geometry", "material"] }],
+    "no-param-reassign": ["error", { "props": true, "ignorePropertyModificationsFor": ["state"] }]
   }
 }
