+++ src/layouts/contexts/NetworkContext.tsx
@@ -1,4 +1,4 @@
-import { getEnvironment } from "@contexts/Environment";
+import { getEnvironment, NetworkConnection } from "@contexts/Environment";
 import { useRouter } from "next/router";
 import { createContext, PropsWithChildren, useContext } from "react";
 import {
@@ -7,15 +7,9 @@ import {
 } from "@defichain/jellyfish-network";
 
 /**
- * What connection is DeFi Scan connected to.
+ * What connection is DeFi Meta Chain connected to.
  * This is different from NetworkName, and should not be used as NetworkName.
  */
-export enum NetworkConnection {
-  LocalPlayground = "Local",
-  RemotePlayground = "Playground",
-  MainNet = "MainNet",
-  TestNet = "TestNet",
-}
 
 export type NetworkName = NetworkObject["name"];
 
@@ -29,16 +23,16 @@ export function useNetwork(): NetworkContextObject {
   return useContext(NetworkContext);
 }
 
-export function NetworkProvider(
-  props: PropsWithChildren<any>
-): JSX.Element | null {
+export function NetworkProvider({
+  children,
+}: PropsWithChildren<any>): JSX.Element | null {
   const router = useRouter();
   const env = getEnvironment();
   const connection = env.resolveConnection(router.query.network);
 
   return (
     <NetworkContext.Provider value={mapNetworkObject(connection)}>
-      {props.children}
+      {children}
     </NetworkContext.Provider>
   );
 }
@@ -46,12 +40,12 @@ export function NetworkProvider(
 function mapNetworkObject(connection: NetworkConnection): NetworkContextObject {
   switch (connection) {
     case NetworkConnection.MainNet:
-      return { connection: connection, ...getNetwork("mainnet") };
+      return { connection, ...getNetwork("mainnet") };
     case NetworkConnection.TestNet:
-      return { connection: connection, ...getNetwork("testnet") };
+      return { connection, ...getNetwork("testnet") };
     case NetworkConnection.RemotePlayground:
     case NetworkConnection.LocalPlayground:
-      return { connection: connection, ...getNetwork("regtest") };
+      return { connection, ...getNetwork("regtest") };
     default:
       throw new Error(`${connection as string} network not found`);
   }
